:recycle: refactor: merge navigation component & devide icons folder
