+++ src/components/icons/ProfileIcon.tsx
@@ -1,5 +1,5 @@
-import type { InternalSvgProps } from '../Svg';
-import Svg from '../Svg';
+import type { InternalSvgProps } from '../common/Svg';
+import Svg from '../common/Svg';
 
 export const ProfileIcon = ({ ...props }: Partial<InternalSvgProps>) => {
   return (
